Move and rename helper/testhelpers/logical to the SDK package
This breaks tests using the original path, and will be fixed in a
follow-up commit
