+++ sdk/acctest/testing_test.go
@@ -1,4 +1,4 @@
-package testing
+package acctest
 
 import (
 	"os"
