+++ sdk/go.mod
@@ -3,31 +3,33 @@ module github.com/hashicorp/vault/sdk
 go 1.13
 
 require (
-	github.com/armon/go-metrics v0.3.0
-	github.com/armon/go-radix v0.0.0-20180808171621-7fddfc383310
-	github.com/aws/aws-sdk-go v1.25.37
+	github.com/armon/go-metrics v0.3.1
+	github.com/armon/go-radix v1.0.0
+	github.com/aws/aws-sdk-go v1.25.41
 	github.com/fatih/structs v1.1.0
 	github.com/go-ldap/ldap/v3 v3.1.3
-	github.com/go-test/deep v1.0.2-0.20181118220953-042da051cf31
+	github.com/go-test/deep v1.0.2
 	github.com/golang/protobuf v1.3.2
 	github.com/golang/snappy v0.0.1
 	github.com/hashicorp/errwrap v1.0.0
 	github.com/hashicorp/go-hclog v0.12.0
-	github.com/hashicorp/go-immutable-radix v1.0.0
+	github.com/hashicorp/go-immutable-radix v1.1.0
 	github.com/hashicorp/go-kms-wrapping/entropy v0.1.0
 	github.com/hashicorp/go-multierror v1.0.0
 	github.com/hashicorp/go-plugin v1.0.1
 	github.com/hashicorp/go-sockaddr v1.0.2
 	github.com/hashicorp/go-uuid v1.0.2
-	github.com/hashicorp/go-version v1.1.0
-	github.com/hashicorp/golang-lru v0.5.1
+	github.com/hashicorp/go-version v1.2.0
+	github.com/hashicorp/golang-lru v0.5.3
 	github.com/hashicorp/hcl v1.0.0
+	github.com/hashicorp/vault v1.4.1
+	github.com/hashicorp/vault/api v1.0.5-0.20200317185738-82f498082f02
 	github.com/mitchellh/copystructure v1.0.0
 	github.com/mitchellh/go-testing-interface v1.0.0
 	github.com/mitchellh/mapstructure v1.1.2
-	github.com/pierrec/lz4 v2.0.5+incompatible
+	github.com/pierrec/lz4 v2.2.6+incompatible
 	github.com/ryanuber/go-glob v1.0.0
-	golang.org/x/crypto v0.0.0-20190418165655-df01cb2cc480
-	golang.org/x/sys v0.0.0-20191008105621-543471e840be
-	google.golang.org/grpc v1.22.0
+	golang.org/x/crypto v0.0.0-20200117160349-530e935923ad
+	golang.org/x/sys v0.0.0-20200122134326-e047566fdf82
+	google.golang.org/grpc v1.23.1
 )
